requirements.txt
astroid==2.0.4
cloudpickle==0.6.1
colorama==0.4.0
cycler==0.10.0
dask==0.20.0
decorator==4.3.0
future==0.17.1
isort==4.3.4
Jinja2==2.10
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
make-deb==0.0.5
MarkupSafe==1.1.0
matplotlib==3.0.2
mccabe==0.6.1
module==0.2.1
networkx==2.2
numpy==1.15.3
opencv-python==3.4.3.18
pbr==5.1.1
Pillow==5.3.0
pylint==2.1.1
pyparsing==2.3.0
python-dateutil==2.7.5
PyWavelets==1.0.1
scikit-image==0.14.1
scipy==1.1.0
six==1.11.0
stevedore==1.30.0
toolz==0.9.0
virtualenv==16.1.0
virtualenv-clone==0.4.0
virtualenvwrapper==4.8.2
wrapt==1.10.11
